# Conflicts:
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/api/common/UserStoreIndexManageResource.kt
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/constant/StoreMessageCode.kt
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/pojo/common/enums/IndexExecuteTimeTypeEnum.kt
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/pojo/common/index/StoreIndexCreateRequest.kt
#	src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/resources/common/UserStoreIndexManageResourceImpl.kt

+++ src/backend/ci/core/quality/api-quality/src/main/kotlin/com/tencent/devops/quality/api/v2/ServiceQualityControlPointMarketResource.kt
@@ -53,7 +53,7 @@ interface ServiceQualityControlPointMarketResource {
         @QueryParam("userId")
         userId: String,
         controlPoint: QualityControlPoint
-    ): Result<Int>
+    ): Result<Long>
 
     @ApiOperation("刷新插件指标的指标")
     @Path("/refreshControlPoint")
